backend/requirements.txt
amqp==5.2.0
asgiref==3.8.1
async-timeout==4.0.3
attrs==23.2.0
beautifulsoup4==4.12.3
billiard==4.2.0
bs4==0.0.2
celery==5.4.0
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
Deprecated==1.2.14
distro==1.9.0
Django==5.0.6
django-admin-sortable2==2.2.1
django-cors-headers==4.4.0
django-debug-toolbar==4.4.2
django-nested-admin==4.0.2
django-rest-passwordreset==1.4.1
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
exceptiongroup==1.2.1
gunicorn==22.0.0
h11==0.14.0
idna==3.7
inflection==0.5.1
kombu==5.3.7
netaddr==1.3.0
outcome==1.3.0.post0
packaging==24.1
prompt_toolkit==3.0.47
psycopg2-binary==2.9.9
PyJWT==2.8.0
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-monkey-business==1.0.0
pytz==2024.1
PyYAML==6.0.1
redis==5.0.7
requests==2.32.3
selenium==4.22.0
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.5
sqlparse==0.5.0
trio==0.26.0
trio-websocket==0.11.1
typing_extensions==4.12.2
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.2
vine==5.1.0
wcwidth==0.2.13
websocket-client==1.8.0
wrapt==1.16.0
wsproto==1.2.0
yookassa==3.3.0
